middleware

+++ routes/web.php
@@ -65,7 +65,7 @@ Route::get('/learnStudent/cards/hiragana', function () {
     return view('student.learn.cards_hiragana',["title" => "learnStudent"]);
 });
 // buat halaman games
-Route::resource('learnStudent/games', GameController::class);
+Route::resource('learnStudent/games', GameController::class)->middleware('auth');
 // buat questions
 // Route::resource('/questions', PointController::class);
 Route::get('/notebook/{unit_id}', [PageController::class, 'showNote']);
@@ -152,59 +152,56 @@ Route::get('/', function () {
 })->middleware('guest');
 Route::get('/homeStudent', function () {
     return view('student.homeStudent3',["title" => "homeStudent"]);
-})->middleware('auth');
+})->middleware('auth')->middleware('student');
 Route::get('/homeTeacher', function () {
     return view('teacher.homeTeacher',["title" => "homeTeacher"]);
-});
+})->middleware('teacher');
 Route::get('/homeAdmin', function () {
     return view('admin.homeAdmin',["title" => "homeAdmin"]);
-});
+})->middleware('admin');
 
 
 //test
 // Article
 
-Route::get('/admin/checkSlug',[ArticleAdminController::class, 'checkSlug']);
-Route::delete('/admin/{article:slug}', [ArticleAdminController::class, 'destroy']);
-Route::get('/admin/{article:slug}/edit', [ArticleAdminController::class, 'edit']);
-Route::put('/admin/{article:slug}', [ArticleAdminController::class, 'update']);
-Route::get('/admin/{article:slug}', [ArticleAdminController::class, 'show']);
-Route::get('/teacher/{article:slug}', [ArticleAdminController::class, 'showteacher']);
-Route::get('/student/{article:slug}', [ArticleAdminController::class, 'showstudent']);
-Route::get('/jpnstudent/{article:id}', [ArticleAdminController::class, 'showstudentjpn']);
-Route::get('/jpnteacher/{article:id}', [ArticleAdminController::class, 'showteacherjpn']);
-Route::get('/jpnadmin/{article:id}', [ArticleAdminController::class, 'jpnshow']);
-Route::get('/admin', [ArticleAdminController::class, 'index']);
-Route::get('/articleStudent', [ArticleAdminController::class, 'indexStudent']);
-Route::get('/articleTeacher', [ArticleAdminController::class, 'indexTeacher']);
-Route::post('/admin', [ArticleAdminController::class, 'store']);
-Route::get('/addArticle', [ArticleAdminController::class, 'create']);
-Route::get('/author/{author:name}',[AuthorController::class, 'index']);
+Route::get('/admin/checkSlug',[ArticleAdminController::class, 'checkSlug'])->middleware('admin');
+Route::delete('/admin/{article:slug}', [ArticleAdminController::class, 'destroy'])->middleware('admin');
+Route::get('/admin/{article:slug}/edit', [ArticleAdminController::class, 'edit'])->middleware('admin');
+Route::put('/admin/{article:slug}', [ArticleAdminController::class, 'update'])->middleware('admin');
+Route::get('/admin/{article:slug}', [ArticleAdminController::class, 'show'])->middleware('admin');
+Route::get('/teacher/{article:slug}', [ArticleAdminController::class, 'showteacher'])->middleware('teacher');
+Route::get('/student/{article:slug}', [ArticleAdminController::class, 'showstudent'])->middleware('student');
+Route::get('/jpnstudent/{article:id}', [ArticleAdminController::class, 'showstudentjpn'])->middleware('student');
+Route::get('/jpnteacher/{article:id}', [ArticleAdminController::class, 'showteacherjpn'])->middleware('teacher');
+Route::get('/jpnadmin/{article:id}', [ArticleAdminController::class, 'jpnshow'])->middleware('admin');
+Route::get('/admin', [ArticleAdminController::class, 'index'])->middleware('admin');
+Route::get('/articleStudent', [ArticleAdminController::class, 'indexStudent'])->middleware('auth')->middleware('student');
+Route::get('/articleTeacher', [ArticleAdminController::class, 'indexTeacher'])->middleware('teacher');
+Route::post('/admin', [ArticleAdminController::class, 'store'])->middleware('admin');
+Route::get('/addArticle', [ArticleAdminController::class, 'create'])->middleware('admin');
+Route::get('/author/{author:name}',[AuthorController::class, 'index'])->middleware('admin');
 
 
 
 // Profile
 
-Route::get('/profileStudent', [StudentController::class, 'index']);
-Route::get('/profileStudent/{user:id}/edit', [StudentController::class, 'edit']);
-Route::put('/profileStudent/{user:id}', [StudentController::class, 'update']);
-Route::delete('/profileStudent/{user:id}', [StudentController::class, 'destroy']);
+Route::get('/profileStudent', [StudentController::class, 'index'])->middleware('auth')->middleware('student');
+Route::get('/profileStudent/{user:id}/edit', [StudentController::class, 'edit'])->middleware('student');
+Route::put('/profileStudent/{user:id}', [StudentController::class, 'update'])->middleware('student');
+Route::delete('/profileStudent/{user:id}', [StudentController::class, 'destroy'])->middleware('student');
 
-Route::get('/profileTeacher', [TeacherController::class, 'index']);
-Route::get('/profileTeacher/{user:id}/edit', [TeacherController::class, 'edit']);
-Route::put('/profileTeacher/{user:id}', [TeacherController::class, 'update']);
-Route::delete('/profileTeacher/{user:id}', [TeacherController::class, 'destroy']);
+Route::get('/profileTeacher', [TeacherController::class, 'index'])->middleware('teacher');
+Route::get('/profileTeacher/{user:id}/edit', [TeacherController::class, 'edit'])->middleware('teacher');
+Route::put('/profileTeacher/{user:id}', [TeacherController::class, 'update'])->middleware('teacher');
+Route::delete('/profileTeacher/{user:id}', [TeacherController::class, 'destroy'])->middleware('teacher');
 
-Route::get('/profileAdmin', [AdminController::class, 'index']);
-Route::get('/profileAdmin/{user:id}/edit', [AdminController::class, 'edit']);
-Route::put('/profileAdmin/{user:id}', [AdminController::class, 'update']);
-Route::delete('/profileAdmin/{user:id}', [AdminController::class, 'destroy']);
+Route::get('/profileAdmin', [AdminController::class, 'index'])->middleware('admin');
+Route::get('/profileAdmin/{user:id}/edit', [AdminController::class, 'edit'])->middleware('admin');
+Route::put('/profileAdmin/{user:id}', [AdminController::class, 'update'])->middleware('admin');
+Route::delete('/profileAdmin/{user:id}', [AdminController::class, 'destroy'])->middleware('admin');
 
 
-Route::get('/teacher', [TeacherController::class, 'showAllTeacher']);
-Route::get('/fetch-cities', [TeacherController::class, 'fetchCities']);
-Route::get('/fetchcities', [TeacherController::class, 'fetchAllCities']);
+Route::get('/teacher', [TeacherController::class, 'showAllTeacher'])->middleware('auth')->middleware('student');
+Route::get('/fetch-cities', [TeacherController::class, 'fetchCities'])->middleware('student');
+Route::get('/fetchcities', [TeacherController::class, 'fetchAllCities'])->middleware('student');
 
-Route::get('/test', function () {
-    return view('student.learn.loadingpage1');
-});
